requirements.txt
fastapi
uvicorn
websockets
openai
langchain
langchain-openai
python-dotenv